Using enum class instead of unsigned int for PointerEventTypes

+++ src/BabylonCpp/src/cameras/inputs/arc_rotate_camera_mouse_wheel_input.cpp
@@ -52,7 +52,7 @@ void ArcRotateCameraMouseWheelInput::attachControl(ICanvas* canvas,
   };
 
   _observer = camera->getScene()->onPointerObservable.add(
-    _wheel, PointerEventTypes::POINTERWHEEL);
+    _wheel, static_cast<int>(PointerEventTypes::POINTERWHEEL));
 }
 
 void ArcRotateCameraMouseWheelInput::detachControl(ICanvas* canvas)
